Create v1 api
In order to create the v1 api for eraser, a number of additional changes
had to be made.

Our eraser-tooling image had to be modified to include the
`conversion-gen` tool. This is to generate type conversions between the
older `v1alpha1` API and the newer `v1`.

To facilitate conversion between the two, an `unversioned` api was
created. This becomes a common hub between `v1` and `v1alpha1` and
facilitates conversion. It also makes it easier for our code: with the
`unversioned` api, we no longer need to update large swaths of code when
a new api version is introduced.

In order to get the `conversion-gen` tool to work, I needed to add
several magic comments throughout the api code. Most importantly, I
needed to add a `doc.go` file to each of the api version directories.
`conversion-gen` looks for this file and will not work without it. Above
the package name in those files, I added the following two magic
comments:

```
+// +k8s:conversion-gen=github.com/Azure/eraser/api/unversioned
+// -external-types=github.com/Azure/eraser/api/v1
```

This tells `conversion-gen` to use the `unversioned` api as the
conversion hub.

To skip kubebuilder processing the `unversioned` api,
`+kubebuilder:skip` had to be added before the package declaration. To
specify which api version should be stored in etcd, the
`+kubebuilder:storageversion` comment was added to the `v1` types. In
addition, `kubebuilder:deprecatedversion:warning="v1alpha1 of the eraser API has been deprecated"`
was added to the `v1alpha1` API in order to warn the user to upgrade.

In the `groupversion_info.go` file of each version directory, I had to
add the line

```go
localSchemeBuilder = runtime.NewSchemeBuilder(SchemeBuilder.AddToScheme)
```

To be perfectly honest, I'm not sure what this does, but it is required
for the `conversion-gen` process to generate working code.

Finally, a new file was added to `test/e2e/test-data` which deploys an
imagelist using the new API version. One of the tests was updated to use
this file instead.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>